refactor(ats-adv): support courant time-step constraint in gwe
