+++ doc/ReleaseNotes/develop.tex
@@ -5,7 +5,7 @@
 	
 %\underline{NEW FUNCTIONALITY}
 %\begin{itemize}
-%	\item xxx
+	\item Support for adjusting time step lengths using the adaptive time stepping (ATS) capability was added to the GWT Advection (ADV) Package of the Groundwater Transport (GWT) Model in release 6.6.0.  The same functionality that was added to GWT is now available with the Groundwater Energy Transport (GWE) Model.  A description of how this functionality works and how to activate it can be found in the release notes for version 6.6.0 (Appendix A) and in the MODFLOW 6 input-output guide.
 %	\item xxx
 %	\item xxx
 %\end{itemize}
